feat: move functions and events inside actors (#582)
### Event Handling Refactoring

The bigger change is we move events emitting out of facade crate into
the respective actors. The facade crate only exposes basic types. It
does not do any functions.

An event now is some domain-specific struct and a corresponding
`TryIntoEVMEvent` trait implementation.

### Code Organisation Improvements

NB. As part of the events change I noticed that we use bunch of
functions like `require_addr_is_origin_or_caller` or `to_id_address`
from `machine` actor in other actors. That leads to an unintended weird
effect: `machine` module from solidity facade crate becomes available
everywhere where the machine actor is declared as a dependency. We have
`recall_actor_sdk` crate, so I moved most of re-used functions from
`machine` actor to this crate.

Every actor that needs to expose an EVM facade now has `sol_facade`
module. Any EVM-related stuff should go there. We talked about events
previously.

As for the function calls, generic structs and traits like
`InvokeContractParams` (incoming EVM message) and `InvokeContractReturn`
went to `recall_actor_sdk`.

We would like to maintain EVM input-output parameters close to the
domain, to the actor code. This input-output parameters translation
could be uniformly implemented as a trait over Solidity call exposed
from the `facade` crate. That means the actor code has to provide
implementation of some trait (same across actors) for structs from
external trait. We introduced `AbiCall` trait for that. It should live
in `recall_actor_sdk` crate. Rust's orphan rule prevent implementing a
foreign trait for a foreign struct (say, implementing `AbiCall` trait
for some struct from `facade` crate inside `blobs` actor) , so we wrap
this `AbiCall` trait into a macro `declare_abi_call`.

### Blobs Functions

Now blobs actor implements two facades: `IBlobsFacade` and
`ICreditFacade` (defined as Solidity interfaces in `facade` crate).

IBlobsFacade:

- addBlob → AddBlob
- deleteBlob → DeleteBlob
- overwriteBlob → OverwriteBlob
- getStats → GetStats
- trimBlobExpiries → TrimBlobExpiries

ICreditFacade:

- buyCredit (for msg.sender and for explicit `recipient`) → BuyCredit
- approveCredit (x3 to allow for different combinations of parameters) →
ApproveCredit
- revokeCredit (x2 to allow for different combinations of parameters) →
RevokeCredit
- setAccountSponsor → SetAccountSponsor
- getAccount → GetAccount
- getCreditApproval → GetCreditApproval
- setAccountStatus → SetAccountSponsor


No Solidity call exposes `from` parameter anymore. We populate it inside
an actor code for the underlying actor call as rt.message.caller(). But,
inside a Rust fn call we still check for "origin or caller" because of
quite convoluted yet relation between bucket and blobs actors.

## Interface changes

getBlob return type on Solidity side now reflects the most recent
changes to get_blob Rust fn.

+++ fendermint/actors/blobs/shared/src/params.rs
@@ -215,7 +215,7 @@ pub struct TrimBlobExpiriesParams {
     pub starting_hash: Option<Hash>,
     /// Limit of blobs to trim expiries for.
     /// This specifies the maximum number of blobs that will be examined for trimming.
-    pub limit: Option<usize>,
+    pub limit: Option<u32>,
 }
 
 /// The stats of the blob actor.
